Add files via upload
This C++ program solves the "Coin Change Problem" using a dynamic programming approach. It calculates the number of ways to make a target sum using an infinite supply of given coin denominations. The solution is efficient and uses a 1D array to optimize space usage.
Time Complexity:  O(N⋅S), where 𝑁 is the number of coins and S is the target sum.
Space Complexity: O(S).