Add files via upload
This C++ program implements the Subset Sum Problem using a Dynamic Programming approach. Given a set of non-negative integers and a target sum, the program checks if there exists a subset of the set whose sum is equal to the given sum. The time complexity of the solution is  O(n×sum), where 𝑛 is the number of elements in the set and sum is the target sum. 
The space complexity is O(n×sum) due to the 2D DP table used.